requirements.txt
dill==0.3.9
GitPython==3.1.43
GitPython==3.1.43
gymnasium==0.29.0
huggingface_hub==0.24.6
hydra-core==1.3.2
matplotlib==3.8.4
mujoco==3.1.6
numpy==1.23.5
omegaconf==2.3.0
omni_isaac_lab==0.17.13
omni_isaac_lab_assets==0.1.3
omni_isaac_lab_tasks==0.7.6
openai==1.55.2
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
Pillow==11.0.0
pycocotools==2.0.8
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scipy==1.14.1
setuptools==69.5.1
stable_baselines3==2.3.2
supervision==0.25.0
torch==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.66.4
transformers==4.44.2
usd_core==23.11
